+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- New command `version` shows the current version of the tool, and of Yarn Spinner itself.
+
 ### Changed
 
 - Updated to [Yarn Spinner v2.3.0](https://github.com/YarnSpinnerTool/YarnSpinner/releases/tag/v2.3.0).